DataChanged (CloudDB) now handles images
Make the DataChanged event consistent with the GotValue event as to how
it handles images.
